Fix the CKF example options (#335)
This fixes a bug in the CKF example command line options where inconsistent type are involved (`int` and `size_t`).

Closes #330.

+++ Examples/Algorithms/TrackFinding/src/TrackFindingOptions.cpp
@@ -19,7 +19,7 @@ void FW::Options::addTrackFindingOptions(FW::Options::Description& desc) {
   auto opt = desc.add_options();
   opt("ckf-slselection-chi2max", value<double>()->default_value(15),
       "Global criteria of maximum chi2 for CKF source link selection");
-  opt("ckf-slselection-nmax", value<int>()->default_value(10),
+  opt("ckf-slselection-nmax", value<size_t>()->default_value(10),
       "Global criteria of maximum number of source link candidates on a "
       "surface for CKF source link selection");
 }
